requirements.txt
pandas
numpy
scikit-learn
nltk
spacy
scipy
fastapi
uvicorn
zenml
mlflow
docker
pytest
httpx
joblib
python-dotenv
SQLAlchemy
psycopg2
beautifulsoup4
xgboost